codegen configuration
configure codegen for the client and server, types generated for the
server, but since the resolvers have not been created for the
client,they have been removed for the client.

codegen was installed at root of the application, with one configuration
file used to determine generated results for both the client and server,
found in the `codegen.ts` file.